// frontend/src/client/types.gen.ts
export type Choice = {
    text: string;
    correct?: boolean;
};
// ...
export type QuestionCreate = {
    question_type: QuestionType;
    question_data: QuestionTeacherData;
    tag?: string;
    total_marks?: number;
    image?: (string | null);
};

export type QuestionStudentResponse = {
    question_type: QuestionType;
    user_response: (string | Array<(string)> | boolean);
};

export type QuestionTeacherData = {
    question_type: QuestionType;
    question_text: string;
    choices?: (Array<Choice> | null);
    true_false_answer?: (boolean | null);
    correct_answer?: (string | null);
};

export type QuestionType = 'single_choice' | 'multiple_choice' | 'user_input' | 'true_false';